% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stock_data.R
\name{plot.stock_data}
\alias{plot.stock_data}
\title{Plot Stock Trend S3 Method}
\usage{
\method{plot}{stock_data}(x, ...)
}
\arguments{
\item{x}{an object of class 'stock_data'.}
}
\description{
A custom visualization method for objects of class 'stock_data'.
}
