% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cln_mrkt_data.R
\name{cln_mrkt_data}
\alias{cln_mrkt_data}
\title{Cleaning Stock Market Data}
\usage{
cln_mrkt_data(data)
}
\arguments{
\item{data}{A data frame or tibble containing stock data.}
}
\value{
A cleaned tibble with the added S3 class 'stock_data'.
}
\description{
This function takes raw messy financial data and standardizes it. It formats date and time as well as standardizes all columns in lowercase. The column names get formatted to remove any underscore or dashes. It also skips any date formatting steps if the data does not have a 'Date' column.
}
